# services/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreatePermissionRequest:
    properties:
      action:
        example: create
        type: string
      description:
        example: Can create new academy
        type: string
      resource:
        example: academy
        type: string
      slug:
        example: academy:create
        type: string
    required:
    - action
    - resource
    type: object
...
  dto.PermissionResponse:
    properties:
      action:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserSimpleResponse'
      deleted_at:
        type: string
      deleted_by:
        $ref: '#/definitions/dto.UserSimpleResponse'
      description:
        type: string
      id:
        type: string
      resource:
        type: string
      slug:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/dto.UserSimpleResponse'
    type: object
...
  dto.UpdatePermissionRequest:
    properties:
      action:
        example: update
        type: string
      description:
        example: Can update academy
        type: string
      resource:
        example: academy
        type: string
      slug:
        example: academy:update
        type: string
    type: object
...
  dto.UserSimpleResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
...
  response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: dev@sportsentral.id
    name: Sport-Sentral Development Team
  description: Centralized API Documentation for Sport-Sentral Microservices Gateway
    (Auth, RBAC, Academy, Venue, Sport, Competition, Scout, Attachment, Log).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sport-Sentral API Gateway
  version: "1.0"
paths:
...
  /admin/permissions:
    post:
      consumes:
      - application/json
      description: Create a new permission entry based on resource and action.
      parameters:
      - description: New Permission Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PermissionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create new permission (Admin)
      tags:
      - Permissions
  /admin/permissions/{id}:
    delete:
      description: Delete permission via soft delete or permanently.
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Permanent delete (default false)
        in: query
        name: permanent
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete permission (Admin)
      tags:
      - Permissions
    put:
      consumes:
      - application/json
      description: Update resource, action, or description of a permission.
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Permission Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PermissionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update permission (Admin)
      tags:
      - Permissions
...
  /permissions:
    get:
      description: Retrieve paginated list of all registered permissions.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PermissionResponse'
                  type: array
              type: object
      summary: Get list of permissions
      tags:
      - Permissions
  /permissions/{id}:
    get:
      description: Retrieve permission data by ID.
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PermissionResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get permission details
      tags:
      - Permissions
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token in the format: `Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"